+++ src/components/fields/input_field.tsx
@@ -11,6 +11,7 @@ interface InputFieldProps {
 	suffix?: string;
 	maxLength?: number;
 	error?: string;
+	"data-testid"?: string;
 }
 
 const InputField: React.FC<InputFieldProps> = ({
@@ -23,11 +24,14 @@ const InputField: React.FC<InputFieldProps> = ({
 	suffix,
 	maxLength,
 	error,
+	"data-testid": dataTestId,
 }) => {
 	return (
-		<div className="mb-4">
+		<div className="">
 			{label && (
-				<label className="block mb-1 text-sm text-gray-700 text-left">{label}</label>
+				<label className="block mb-1 text-sm text-gray-700 text-left">
+					{label}
+				</label>
 			)}
 
 			<div className={`flex ${error ? "flex-col" : ""}`}>
@@ -38,6 +42,7 @@ const InputField: React.FC<InputFieldProps> = ({
 						value={value}
 						onChange={onChange}
 						maxLength={maxLength}
+						data-testid={dataTestId}
 						className={`px-3 py-2 border focus:outline-none focus:ring ${
 							error
 								? "border-red-500 focus:ring-red-200"